refactor: apply PR feedback to improve prompts module
Address code review suggestions:
- Use barrel export consistently across all imports
- Consolidate PromptManager and GeminiPrompts imports in main.ts
- Add declarations.d.ts for proper .txt module typing
- Remove @ts-ignore comments from gemini-prompts.ts

These changes improve code maintainability and follow TypeScript
best practices for module declarations.
