+++ src/completions.ts
@@ -2,7 +2,7 @@ import ObsidianGemini from './main';
 import { MarkdownView, debounce, Notice } from 'obsidian';
 import { forceableInlineSuggestion, Suggestion } from 'codemirror-companion-extension';
 import { BaseModelRequest } from './api/index';
-import { GeminiPrompts } from './prompts/gemini-prompts';
+import { GeminiPrompts } from './prompts';
 import { GeminiClientFactory } from './api/simple-factory';
 
 export class GeminiCompletions {
