+++ src/files/index.ts
@@ -1,7 +1,7 @@
 import ObsidianGemini from '../main';
 import { TFile } from 'obsidian';
 import { logDebugInfo } from '../api/utils/debug';
-import { GeminiPrompts } from '../prompts/gemini-prompts';
+import { GeminiPrompts } from '../prompts';
 
 export class ScribeFile {
 	private plugin: ObsidianGemini;
