+++ src/prompts/gemini-prompts.ts
@@ -2,23 +2,14 @@ import * as Handlebars from 'handlebars';
 import { CustomPrompt } from './types';
 import ObsidianGemini from '../main';
 
-// @ts-ignore
 import systemPromptContent from '../../prompts/systemPrompt.txt';
-// @ts-ignore
 import completionPromptContent from '../../prompts/completionPrompt.txt';
-// @ts-ignore
 import generalPromptContent from '../../prompts/generalPrompt.txt';
-// @ts-ignore
 import summaryPromptContent from '../../prompts/summaryPrompt.txt';
-// @ts-ignore
 import contextPromptContent from '../../prompts/contextPrompt.txt';
-// @ts-ignore
 import selectionRewritePromptContent from '../../prompts/selectionRewritePrompt.txt';
-// @ts-ignore
 import agentToolsPromptContent from '../../prompts/agentToolsPrompt.txt';
-// @ts-ignore
 import vaultAnalysisPromptContent from '../../prompts/vaultAnalysisPrompt.txt';
-// @ts-ignore
 import imagePromptGeneratorContent from '../../prompts/imagePromptGenerator.txt';
 
 export class GeminiPrompts {
