+++ src/rewrite-selection.ts
@@ -1,7 +1,7 @@
 import ObsidianGemini from './main';
 import { Editor, Notice } from 'obsidian';
 import { ExtendedModelRequest } from './api/index';
-import { GeminiPrompts } from './prompts/gemini-prompts';
+import { GeminiPrompts } from './prompts';
 import { GeminiClientFactory } from './api/simple-factory';
 
 export class SelectionRewriter {
